requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib
bcrypt
python-multipart
aiofiles
python-jose